mcproxy: Wan port does not sent multicst listener report
Wan port does not sent multicst listener report for two sources
specified

This fixes the wrong checksum algorithm issue and a typo in the
mld source filter file.

Add mcproxy hotplug script when the wan connection is reestablished
Set the default upstream and downstream when they are unset.
the default upstream is the wan interface
the default downstream is the lan interface.

Change-Id: Ifc622d45ef1eba2e0dcca060bcc4b8854243e678
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>