nat46: fix ICMPv6 error message dropped locally
While routing IPv6 packets from a customer-side translated device (CLAT)
to a provider-side translated device (PLAT), it is possible that the IPv6
destination is unknown. In such a scenario, the IPv6 stack must send back
an ICMP error. However, the source IPv6 address of this error message does
not have a MAP-T translation. According to RFC2473, the translation layer
should use the tunnel's own IPv4 address for the IPv6 ICMP packet's source
address.

Change-Id: I280724d1d06cf0de586d13c82b46f1e3ab8cead4
Signed-off-by: Ken Zhu <guigenz@codeaurora.org>